+++ mkdocs.yml
@@ -17,6 +17,7 @@ nav:
           - 'Embeddings: 2D and Ambient': 'map_state/embeddings.md'
           - Atlas Topics: 'map_state/topics.md'
           - Atlas Semantic Clusters: 'map_state/duplicates.md'
+          - Atlas Tags: 'map_state/tags.md'
 
       - 'atlas_api.md'
