+++ nomic/data_operations.py
@@ -14,7 +14,7 @@ import pandas as pd
 import pyarrow as pa
 import requests
 from loguru import logger
-from pyarrow import compute as pc
+from pyarrow import compute as pc, ipc
 from tqdm import tqdm
 
 from .settings import EMBEDDING_PAGINATION_LIMIT
@@ -460,7 +460,7 @@ class AtlasMapEmbeddings:
             if check_access:
                 return
 
-            shard_name = '{}_{}_{}.feather'.format(self.atlas_index_id, offset, offset + limit)
+            shard_name = '{}_{}_{}.feather'.format(self.project.atlas_index_id, offset, offset + limit)
             shard_path = os.path.join(save_directory, shard_name)
             try:
 
@@ -492,3 +492,106 @@ class AtlasMapEmbeddings:
 
     def __repr__(self) -> str:
         raise NotImplementedError()
+
+
+class AtlasMapTags:
+    """
+    Atlas Map Tag State
+
+    Tags are shared acrossed all maps in your AtlasProject.
+    """
+
+    def __init__(self, projection: "AtlasProjection"):
+        self.projection = projection
+        self.project = projection.project
+        self.id_field = self.projection.project.id_field
+
+    def get_tags(self) -> Dict[str, List[str]]:
+        '''
+        Retrieves back all tags made in the web browser for a specific project and map.
+
+        Returns:
+            A dictionary mapping data points to tags.
+        '''
+        # now get the tags
+        datums_and_tags = requests.post(
+            self.project.atlas_api_path + '/v1/project/tag/read/all_by_datum',
+            headers=self.project.header,
+            json={
+                'project_id': self.project.id,
+            },
+        ).json()['results']
+
+        label_to_datums = {}
+        for item in datums_and_tags:
+            for label in item['labels']:
+                if label not in label_to_datums:
+                    label_to_datums[label] = []
+                label_to_datums[label].append(item['datum_id'])
+        return label_to_datums
+
+    def tag(self, ids: List[str], tags: List[str]):
+        '''
+        Adds tags to datapoints.
+
+        Args:
+            ids: The datum ids you want to tag
+            tags: A list containing the tags you want to apply to these data points.
+
+        '''
+        assert isinstance(ids, list), 'ids must be a list of strings'
+        assert isinstance(tags, list), 'tags must be a list of strings'
+
+        colname = json.dumps(
+            {'project_id': self.project.id, 'atlas_index_id': self.projection.atlas_index_id, 'type': 'datum_id', 'tags': tags}
+        )
+        payload_table = pa.table([pa.array(ids, type=pa.string())], [colname])
+        buffer = io.BytesIO()
+        writer = ipc.new_file(buffer, payload_table.schema, options=ipc.IpcWriteOptions(compression='zstd'))
+        writer.write_table(payload_table)
+        writer.close()
+        payload = buffer.getvalue()
+
+        headers = self.project.header.copy()
+        headers['Content-Type'] = 'application/octet-stream'
+        response = requests.post(self.project.atlas_api_path + "/v1/project/tag/add", headers=headers, data=payload)
+        if response.status_code != 200:
+            raise Exception("Failed to add tags")
+
+    def remove_tags(self, ids: List[str], tags: List[str], delete_all: bool = False) -> bool:
+        '''
+        Deletes the specified tags from the given data points.
+
+        Args:
+            ids: The datum_ids to delete tags from.
+            tags: The list of tags to delete from the data points. Each tag will be applied to all data points in `ids`.
+            delete_all: If true, ignores datum_ids and deletes all specified tags from all data points.
+
+        Returns:
+            True on success
+
+        '''
+        assert isinstance(ids, list), 'datum_ids must be a list of strings'
+        assert isinstance(tags, list), 'tags must be a list of strings'
+
+        colname = json.dumps(
+            {
+                'project_id': self.project.id,
+                'atlas_index_id': self.projection.atlas_index_id,
+                'type': 'datum_id',
+                'tags': tags,
+                'delete_all': delete_all,
+            }
+        )
+        payload_table = pa.table([pa.array(ids, type=pa.string())], [colname])
+        buffer = io.BytesIO()
+        writer = ipc.new_file(buffer, payload_table.schema, options=ipc.IpcWriteOptions(compression='zstd'))
+        writer.write_table(payload_table)
+        writer.close()
+        payload = buffer.getvalue()
+
+        headers = self.project.header.copy()
+        headers['Content-Type'] = 'application/octet-stream'
+        response = requests.post(self.project.atlas_api_path + "/v1/project/tag/delete", headers=headers, data=payload)
+        if response.status_code != 200:
+            raise Exception("Failed to delete tags")
