+++ nomic/project.py
@@ -34,7 +34,7 @@ import nomic
 
 from .cli import refresh_bearer_token, validate_api_http_response
 from .data_inference import convert_pyarrow_schema_for_atlas
-from .data_operations import AtlasMapDuplicates, AtlasMapEmbeddings, AtlasMapTopics
+from .data_operations import AtlasMapDuplicates, AtlasMapEmbeddings, AtlasMapTopics, AtlasMapTags
 from .settings import *
 from .utils import assert_valid_project_id, get_object_size_in_bytes
 
@@ -494,6 +494,13 @@ class AtlasProjection:
             self._embeddings = AtlasMapEmbeddings(self)
         return self._embeddings
 
+    @property
+    def tags(self):
+        """Embedding state"""
+        if self._tags is None:
+            self._tags = AtlasMapTags(self)
+        return self._tags
+
     def _fetch_tiles(self, overwrite: bool = True):
         """
         Downloads all web data for the projection to the specified directory and returns it as a memmapped arrow table.
@@ -605,95 +612,6 @@ class AtlasProjection:
             return response.json()['atoms']
         else:
             raise Exception(response.text)
-
-    def get_tags(self) -> Dict[str, List[str]]:
-        '''
-        Retrieves back all tags made in the web browser for a specific project and map.
-
-        Returns:
-            A dictionary mapping datum ids to tags.
-        '''
-        # now get the tags
-        datums_and_tags = requests.post(
-            self.project.atlas_api_path + '/v1/project/tag/read/all_by_datum',
-            headers=self.project.header,
-            json={
-                'project_id': self.project.id,
-            },
-        ).json()['results']
-
-        label_to_datums = {}
-        for item in datums_and_tags:
-            for label in item['labels']:
-                if label not in label_to_datums:
-                    label_to_datums[label] = []
-                label_to_datums[label].append(item['datum_id'])
-        return label_to_datums
-
-    def tag(self, ids: List[str], tags: List[str]):
-        '''
-
-        Args:
-            ids: The datum ids you want to tag
-            tags: A list containing the tags you want to apply to these data points.
-
-        '''
-        assert isinstance(ids, list), 'ids must be a list of strings'
-        assert isinstance(tags, list), 'tags must be a list of strings'
-
-        colname = json.dumps(
-            {'project_id': self.project.id, 'atlas_index_id': self.atlas_index_id, 'type': 'datum_id', 'tags': tags}
-        )
-        payload_table = pa.table([pa.array(ids, type=pa.string())], [colname])
-        buffer = io.BytesIO()
-        writer = ipc.new_file(buffer, payload_table.schema, options=ipc.IpcWriteOptions(compression='zstd'))
-        writer.write_table(payload_table)
-        writer.close()
-        payload = buffer.getvalue()
-
-        headers = self.project.header.copy()
-        headers['Content-Type'] = 'application/octet-stream'
-        response = requests.post(self.project.atlas_api_path + "/v1/project/tag/add", headers=headers, data=payload)
-        if response.status_code != 200:
-            raise Exception("Failed to add tags")
-
-    def remove_tags(self, ids: List[str], tags: List[str], delete_all: bool = False) -> bool:
-        '''
-        Deletes the specified tags from the given datum_ids.
-
-        Args:
-            ids: The datum_ids to delete tags from.
-            tags: The list of tags to delete from the data points. Each tag will be applied to all data points in `ids`.
-            delete_all: If true, ignores datum_ids and deletes all specified tags from all data points.
-
-        Returns:
-            True on success
-
-        '''
-        assert isinstance(ids, list), 'datum_ids must be a list of strings'
-        assert isinstance(tags, list), 'tags must be a list of strings'
-
-        colname = json.dumps(
-            {
-                'project_id': self.project.id,
-                'atlas_index_id': self.atlas_index_id,
-                'type': 'datum_id',
-                'tags': tags,
-                'delete_all': delete_all,
-            }
-        )
-        payload_table = pa.table([pa.array(ids, type=pa.string())], [colname])
-        buffer = io.BytesIO()
-        writer = ipc.new_file(buffer, payload_table.schema, options=ipc.IpcWriteOptions(compression='zstd'))
-        writer.write_table(payload_table)
-        writer.close()
-        payload = buffer.getvalue()
-
-        headers = self.project.header.copy()
-        headers['Content-Type'] = 'application/octet-stream'
-        response = requests.post(self.project.atlas_api_path + "/v1/project/tag/delete", headers=headers, data=payload)
-        if response.status_code != 200:
-            raise Exception("Failed to delete tags")
 
 
 class AtlasProject(AtlasClass):
